backend/requirements.txt
Flask>=2.2
flask-cors
flask-jwt-extended
flask-bcrypt
mysql-connector-python
PyMySQL
